Add is_detected and nonesuch
Tests for nonesuch to follow later, since it requires several traits
(e.g is_copy_assignable to ensure that it is not copyable) whose
implementation require is_detected. It's OK to postpone this testing
since is_detected uses nonesuch as a tag type, so there is no dependence
on its behavior.