Fix is_function instantiating more types than necessary through ADL
is_function was invoking ADL when checking for references to functions
via a call to decay_check. This creates problems when ADL tries to look
up extra "related" types, such as the type being pointed to when a
pointer to an uninstantiable type is passed as an argument.

This commit suppresses ADL by qualifying the call to decay_check. As a
quick lint, it also replaces the referenceability check via is_detected
with is_referenceable.