Update alignment_of's comments to indicate when program is ill-formed
This commit adds a note to mention that the program will not compile
when T's alignment will overflow ptrdiff_t. This is due to the use of
size_to_ptrdiff when casting alignas' result to ptrdiff_t; a
static_assert checks for overflow. In reality, this overflow should
probably not happen - I think compilers should error out on any
alignment requirement approaching PTRDIFF_MAX.