Fix missing conversion in one of common_type's test cases
The test case for user-provided common_type specializations names
user_spec<T> as the common type between user_spec<T> and T. However,
user_spec<T> itself is not convertible from T, so it doesn't make sense
for it to be the common type. This commits enables explicit conversion
from T to user_spec<T> so that the test case makes sense.