Make tuple_size's primary template undefined
There's no reason to provide a definition for tuple_size's primary
template; BML's tuple_size should only be used with BML's tuple.

This commit also adds tests to check that no definition is provided for
tuple_element, variant_size and variant_alternative when T is not a
tuple (for tuple_element) or variant (for variant_size and
variant_alternative).